fix: resolve build errors and architectural violations in dual-mode search
- Fixed 'm_activeSearchId' undeclared identifier by correctly placing it in ContentPanel.h.
- Corrected std::vector usage from isEmpty() to empty() in CoreController.cpp.
- Unified ItemRecord creation logic into IndexedEntry to decouple UI and Core layers.
- Resolved multiple symbol definition and visibility issues in ContentPanel.
- Ensured thread safety for searchId management using std::atomic.
- Verified context-aware switching logic between Categories and NavPanel.

+++ src/core/CoreController.cpp
@@ -95,7 +95,7 @@ QStringList CoreController::performSearch(const QString& keyword, const QString&
                 if (count >= 10000) break;
             }
 
-            if (!batch.empty() && weakThis && weakThis->m_currentSearchId == searchId) {
+            if (!batch.empty() && weakThis && (int)weakThis->m_currentSearchId == searchId) {
                 QMetaObject::invokeMethod(weakThis.data(), "searchResultFound", Qt::QueuedConnection, Q_ARG(int, searchId), Q_ARG(std::vector<ArcMeta::ItemRecord>, batch));
             }
             if (weakThis && weakThis->m_currentSearchId == searchId) {
